add in more stuff:

+++ program/src/language_equivalences.ml
@@ -55,8 +55,8 @@ and dnf_regex_to_regex (r:dnf_regex) : regex =
 
 
 type queue_element =
-  | QERegexCombo of regex * regex * float * int
-  | QEGenerator of (unit -> ((queue_element * float) list))
+  | QERegexCombo of regex * regex * int * int
+  | QEGenerator of (unit -> ((queue_element * int) list))
 
 
 let clause_to_kvp ((atoms,strings):clause)
